\name{calcIntersect}
\alias{calcIntersect}
\title{Calculates the position of a point along a segment.}
\usage{
  calcIntersect(z1, z2, z)
}
\arguments{
  \item{z1}{First point.}

  \item{z2}{Second point.}

  \item{z}{Point whose position is to be determined.}
}
\value{
  The linear position of the point where z1 is 0 and z2 is
  1.
}
\description{
  Thus function calculates the position along a segment
  where a point is located. The position is non-dimensional
  in the sense that the first point is at position 0 and
  the second point is at position 1.
}

