\name{faceProject2d.default}
\alias{faceProject2d.default}
\title{Method that projects a 3D points  into a plane.}
\usage{
  faceProject2d.default(f, base2d)
}
\arguments{
  \item{f}{Matrix where each row is a 3D point}

  \item{base2d}{Orthonormal local basis}
}
\value{
  2D projection
}
\description{
  Method that projects a 3D points into a plane.
}

