\name{faceProject2d.mesh3d}
\alias{faceProject2d.mesh3d}
\title{Method that projects a 3D triangular mesh  into a plane.}
\usage{
  faceProject2d.mesh3d(f, base2d)
}
\arguments{
  \item{f}{mesh3d instance}

  \item{base2d}{Orthonormal local basis}
}
\value{
  2D projection, mesh instance
}
\description{
  Method that projects a 3D triangular mesh into a plane.
}

