\name{faceProject2d.polygon3d}
\alias{faceProject2d.polygon3d}
\title{Method that projects a 3D polygon  into a plane.}
\usage{
  faceProject2d.polygon3d(f, base2d)
}
\arguments{
  \item{f}{polygon3d instance}

  \item{base2d}{Orthonormal local basis}
}
\value{
  2D projection, polygon
}
\description{
  Method that projects a 3D polygon into a plane.
}

