\name{polygonCentroid.polygon3d}
\alias{polygonCentroid.polygon3d}
\title{Centroid of a 3D polygon.}
\usage{
  polygonCentroid.polygon3d(p)
}
\arguments{
  \item{p}{Polygon}
}
\value{
  Coordinates of the centroid
}
\description{
  Method that computes centroid of a polygons of class
  polygon3d.
}

