\name{polygonIntersect.gpc.poly}
\alias{polygonIntersect.gpc.poly}
\title{Polygon Intersection.}
\usage{
  polygonIntersect.gpc.poly(p1, p2)
}
\description{
  Method that computes intersection of two gpc.poly
  polygons. Actually calls the approppriate function of
  gpclib to compute the intersection.
}

