\name{polygonIntersect}
\alias{polygonIntersect}
\title{Polygon Intersection.}
\usage{
  polygonIntersect(p1, p2)
}
\arguments{
  \item{p1}{First polygon}

  \item{p2}{Second polygon}
}
\value{
  The intersection of the polygons.
}
\description{
  S3 generic function that compute the intersection of two
  polygons. Remember that the intersection of two polygons
  can be 0, 1 or more polygons.
}

