\name{polygonOrient.default}
\alias{polygonOrient.default}
\title{Default method for polygon orientation.}
\usage{
  polygonOrient.default(p)
}
\description{
  The first row should be x coordinates. second row should
  be y coordinates. If a third row is present this function
  will use the 3D method.
}

