\name{polygonOrient.polygon3d}
\alias{polygonOrient.polygon3d}
\title{Method for 3D polygon orientation.}
\usage{
  polygonOrient.polygon3d(p)
}
\description{
  The function returns the normal to the polygon according
  to right hand rule orientation. If the area is zero,
  return the null vector c(0,0,0).
}

