\name{project2dTo3d.default}
\alias{project2dTo3d.default}
\title{Get back 3D points from 2D projection.}
\usage{
  project2dTo3d.default(p, pinfo)
}
\arguments{
  \item{p}{2D projection of points}

  \item{pinfo}{Projection info obtained from
  \code{\link{projectionInfo}}}
}
\value{
  3D points.
}
\description{
  Retrieves initial 3D points from ist 2D projection.
}
\seealso{
  \code{\link{projectionInfo}} ,
  \code{\link{project3dTo2d}}
}

