\name{project2dTo3d}
\alias{project2dTo3d}
\title{Get back 3D data from 2D projection.}
\usage{
  project2dTo3d(p, pinfo)
}
\arguments{
  \item{p}{2D projection}

  \item{pinfo}{Projection info obtained from
  \code{\link{projectionInfo}}}
}
\value{
  3D information.
}
\description{
  Retrieves initial 3D data from the 2D projection.
}
\seealso{
  \code{\link{projectionInfo}}, \code{\link{project3dTo2d}}
}

