\name{project3dTo2d.pmesh3d}
\alias{project3dTo2d.pmesh3d}
\title{Method that projects a 3D polygon mesh  into a plane.}
\usage{
  project3dTo2d.pmesh3d(p, base2d)
}
\arguments{
  \item{f}{pmesh3d instance}

  \item{base2d}{Orthonormal local basis}
}
\value{
  2D projection, pmesh instance
}
\description{
  Method that projects a 3D polygon mesh into a plane.
}

