\name{project3dTo2d.polygon3d}
\alias{project3dTo2d.polygon3d}
\title{Method that projects a 3D polygon  into a plane.}
\usage{
  project3dTo2d.polygon3d(p, base2d)
}
\arguments{
  \item{f}{polygon3d instance}

  \item{base2d}{Orthonormal local basis}
}
\value{
  2D projection, polygon
}
\description{
  Method that projects a 3D polygon into a plane.
}
\seealso{
  \code{\link{faceProject2d}}, \code{\link{polygonProject}}
}

